+++ gdsfactory/components/text_rectangular.py
@@ -1,44 +1,82 @@
-from typing import Tuple
+from typing import Callable, Tuple
 
 import gdsfactory as gf
 from gdsfactory.component import Component
-from gdsfactory.components.manhattan_font import manhattan_text
-from gdsfactory.types import Coordinate, Layer
+from gdsfactory.components.copy_layers import copy_layers
+from gdsfactory.components.text_rectangular_font import pixel_array, rectangular_font
+from gdsfactory.tech import LAYER
 
 
 @gf.cell
 def text_rectangular(
     text: str = "abcd",
     size: float = 10.0,
-    position: Coordinate = (0, 0),
+    position: Tuple[float, float] = (0.0, 0.0),
     justify: str = "left",
-    layers: Tuple[Layer, ...] = ((1, 0),),
+    layer: Tuple[int, int] = (1, 0),
+    font: Callable = rectangular_font,
 ) -> Component:
-    """Returns Pixel based rectangular text.
+    """Pixel based font, guaranteed to be manhattan, without acute angles.
 
     Args:
-        text: text
+        text:
         size: pixel size
         position: coordinate
-        justify: left, right, center
-        layers: layers
-    """
-    c = gf.Component()
-    for layer in layers:
-        c.add_ref(
-            manhattan_text(
-                text=text, size=size, position=position, justify=justify, layer=layer
-            )
-        )
-    return c
+        justify: left, right or center
+        layer:
+        font: function to get the dictionary of characters
 
+    """
+    pixel_size = size
+    xoffset = position[0]
+    yoffset = position[1]
+    component = gf.Component()
+    characters = rectangular_font()
+
+    for line in text.split("\n"):
+        for character in line:
+            if character == " ":
+                xoffset += pixel_size * 6
+            elif character.upper() not in characters:
+                print(f"skipping character {character} not in font")
+            else:
+                pixels = characters[character.upper()]
+                ref = component.add_ref(
+                    pixel_array(pixels=pixels, pixel_size=pixel_size, layer=layer)
+                )
+                ref.move((xoffset, yoffset))
+                component.absorb(ref)
+                xoffset += pixel_size * 6
+
+        yoffset -= pixel_size * 6
+        xoffset = position[0]
+    justify = justify.lower()
+    for ref in component.references:
+        if justify == "left":
+            pass
+        elif justify == "right":
+            ref.xmax = position[0]
+        elif justify == "center":
+            ref.move(origin=ref.center, destination=position, axis="x")
+        else:
+            raise ValueError(f"justify = {justify} not valid (left, center, right)")
+
+    return component
+
+
+text_rectangular_multi_layer = gf.partial(
+    copy_layers,
+    factory=text_rectangular,
+    layers=(LAYER.WG, LAYER.M1, LAYER.M2, LAYER.M3),
+)
 
 if __name__ == "__main__":
-    c = text_rectangular(
-        text=".[,ABCDEFGHIKKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789:/",
-        size=4.0,
-        justify="right",
-        position=(120.5, 3),
-        layers=[(1, 0), (2, 0)],
+    import string
+
+    c = text_rectangular_multi_layer(
+        # text="The mask is nearly done. only 12345 drc errors remaining?",
+        # text="v",
+        text=string.ascii_lowercase,
+        layers=(LAYER.SLAB90, LAYER.M2),
     )
     c.show()